+++ gimmemotifs/motif/_scanning.py
@@ -5,6 +5,8 @@
 # distribution.
 """Scanning functions for Motif class"""
 
+from warnings import warn
+
 from gimmemotifs.c_metrics import pfmscan, pwmscan
